+++ crates/module-system/module-implementations/sov-sequencer-registry/src/lib.rs
@@ -32,14 +32,6 @@ use thiserror::Error;
 )]
 #[serde(rename_all = "snake_case")]
 pub enum AllowedSequencerError {
-    /// The amount of gas tokens that the sender is has staken is too low.
-    #[error("The amount staked by the sequencer is less than the minimum bond. Amount currently staked: {bond_amount}, minimum bond amount: {minimum_bond_amount}.")]
-    InsufficientStakeAmount {
-        /// The amount of gas tokens that the sender is has staken.
-        bond_amount: Amount,
-        /// The minimum amount of gas tokens that the sequencer must stake.
-        minimum_bond_amount: Amount,
-    },
     /// The sequencer is not registered.
     #[error("The sequencer is not registered.")]
     NotRegistered,