+++ crates/module-system/sov-modules-stf-blueprint/src/proof_processing.rs
@@ -1,8 +1,7 @@
 use std::marker::PhantomData;
 
 use sov_modules_api::capabilities::{
-    AuthorizeSequencerError, GasEnforcer, ProofProcessor, SequencerAuthorization,
-    SequencerRemuneration, TryReserveGasError,
+    GasEnforcer, ProofProcessor, SequencerAuthorization, SequencerRemuneration, TryReserveGasError,
 };
 use sov_modules_api::proof_metadata::{ProofType, SerializeProofWithDetails};
 use sov_modules_api::transaction::AuthenticatedTransactionData;
@@ -38,29 +37,15 @@ where
 {
     let workflow = ProofProcessingWorkflow::new(runtime, blob_hash, &sequencer_da_address);
 
-    let max_tx_check_costs = <S as GasSpec>::max_tx_check_costs();
-
     // Check if the sequencer is bonded, and create `pre_exec_working_set`.
     let (sequencer_rollup_address, mut pre_exec_working_set) =
-        match workflow.authorize_sequencer(gas_price, max_tx_check_costs, state.to_tx_scratchpad())
-        {
+        match workflow.authorize_sequencer(gas_price, state.to_tx_scratchpad()) {
             WorkflowResult::Proceed(pre_exec_working_set) => pre_exec_working_set,
             WorkflowResult::EarlyReturn(out, state) => {
                 tracing::debug!("{LOG_PREFIX}: unable to create pre execution working set");
                 return (out, state);
             }
         };
-
-    // This represents the cost incurred by the sequencer solely for accepting the proof. It includes the cost of:
-    // - refund_remaining_gas
-    // - reward_sequencer
-    // etc
-    let tx_precessing_costs = <S as GasSpec>::process_tx_pre_exec_checks_gas();
-
-    // It is ok to panic here because we asserter that the sequencer has enough funds to pay for the tx processing costs.
-    pre_exec_working_set
-        .charge_gas(&tx_precessing_costs)
-        .expect("The rollup is misconfigured: MAX_SEQUENCER_EXEC_GAS_PER_TX: {max_tx_check_costs} must be greater than PROCESS_TX_PRE_EXEC_GAS: {tx_precessing_costs}.");
 
     match SerializeProofWithDetails::<S>::deserialize(
         &mut raw_proof.as_slice(),
@@ -162,9 +147,9 @@ where
                 scratchpad.commit(),
             )
         }
-        Err(_) => {
+        Err(e) => {
             // We could not deserialize the data from the DA. Penalize the sequencer and return early.
-            tracing::debug!("{LOG_PREFIX}: unable to deserialize proof");
+            tracing::debug!("{LOG_PREFIX}: unable to deserialize proof {:?}", e);
 
             let (state, gas_meter) = pre_exec_working_set.to_scratchpad_and_gas_meter();
             let gas_used = gas_meter.gas_info().gas_used;
@@ -241,23 +226,26 @@ where
     fn authorize_sequencer<I: StateProvider<S>>(
         &self,
         gas_price: &<S::Gas as Gas>::Price,
-        max_tx_check_costs: S::Gas,
         mut tx_scratchpad: TxScratchpad<S, I>,
     ) -> PreExecWorkingSetResult<S, I> {
-        let max_auth_cost = max_tx_check_costs.value(gas_price);
-        match self.runtime.sequencer_authorization().authorize_sequencer(
-            self.sequencer_da_address,
-            max_auth_cost,
-            &mut tx_scratchpad,
-        ) {
-            Ok(allowed_sequencer) => {
-                let gas_meter =
-                    BasicGasMeter::<S::Gas>::new_with_gas(max_tx_check_costs, gas_price.clone());
-                let pre_exec_working_set = tx_scratchpad.to_pre_exec_working_set(gas_meter);
+        let sequencer = self
+            .runtime
+            .sequencer_authorization()
+            .authorize_sequencer(self.sequencer_da_address, &mut tx_scratchpad)
+            .expect("Blob selection must guarantee that sequencer is registered");
 
-                WorkflowResult::Proceed((allowed_sequencer.address, pre_exec_working_set))
-            }
-            Err(AuthorizeSequencerError { reason }) => WorkflowResult::EarlyReturn(
+        let gas_meter = BasicGasMeter::<S::Gas>::new(sequencer.balance, gas_price.clone());
+        let mut pre_exec_working_set = tx_scratchpad.to_pre_exec_working_set(gas_meter);
+
+        // This represents the cost incurred by the sequencer solely for accepting the proof. It includes the cost of:
+        // - refund_remaining_gas
+        // - reward_sequencer
+        // etc
+        let tx_preprocessing_costs = <S as GasSpec>::process_tx_pre_exec_checks_gas();
+
+        match pre_exec_working_set.charge_gas(&tx_preprocessing_costs) {
+            Ok(_) => WorkflowResult::Proceed((sequencer.address, pre_exec_working_set)),
+            Err(reason) => WorkflowResult::EarlyReturn(
                 ProcessProofOutput {
                     proof_receipt: invalid_proof_receipt::<S>(
                         self.blob_hash,
@@ -268,7 +256,10 @@ where
                     ),
                     gas_used: S::Gas::zero(),
                 },
-                tx_scratchpad.commit(),
+                pre_exec_working_set
+                    .to_scratchpad_and_gas_meter()
+                    .0
+                    .commit(),
             ),
         }
     }
@@ -352,7 +343,7 @@ where
                 self.sequencer_da_address,
                 &mut state,
             )
-            // We ensured this before we started processing the proof.
+            // This **should** never fail because we initialize gas meter with the sequencer bond and this should account for the reward costs. If not that is a bug and we should panic.
             .expect("Sequencer should have enough funds to pay for the penalty");
 
         state